requirements.txt
pytorch >= 1.11.0
torchvision
matplotlib >= 3.7
scikit-learn >= 1.0
ipykernel
jupyter
timm
